check rmdir to be non-empty, check if rm is not directory

+++ src/commandows/Rm.java
@@ -2,16 +2,35 @@ package commandows;
 
 import java.io.File;
 
+import wandows.Main;
+
 public class Rm {
 	public Rm(String[] args) {
 		try {
-			File file = new File(args[0]);
-			if (file.exists()) {
-				file.delete();
+			if(args.length == 0) {
+				Main.outln("rm: please specify a file");
+				return;
 			}
+			
+			File file = new File(args[0]);
+			String dir = file.getPath();
+			String fname = file.getName();
+			
+			if(!dir.equals(Main.getCurrentWorkingDirectory()))
+				dir = Main.getCurrentWorkingDirectory();
+			
+			if(!file.exists() && !file.isDirectory()) {
+				file = new File(dir + "\\" + fname);
+				
+				// if it doesn't exist, throw an error
+				if(file.exists())
+					file.delete();
+				else
+					Main.outln("rmdir: `" + file.getCanonicalFile() + "`: file not found or is a directory");
+			} else
+				Main.outln("rmdir: `" + file.getCanonicalFile() + "`: file not found or is a directory");
 		} catch (Exception e) {
 			System.err.println(e.getMessage());
 		}
 	}
-
 }
